requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
pyvesync==2.1.14
python-multipart==0.0.6
typing-extensions==4.8.0
pytest==7.4.3 
